iot_edge/humid_telemetry/modules/ml_inference/requirements.txt
numpy
python-multipart
scikit-learn
joblib
azure-iot-device